# resources/openapi.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Location Tracker API
servers:
  - url: http://localhost:3000/
paths:
...
  /users/auth:
    post:
      tags:
        - Users
      description: Authenticate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuth'
      responses:
        202:
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token
  /users/register:
    post:
      tags:
        - Users
      description: Register new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        201:
          description: New user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
...
    User:
      description: User
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: User id
        alias:
          type: string
          minLength: 1
          maxLength: 512
          description: User alias
        publicKey:
          type: string
          description: User public key
    UserCreate:
      description: User create data
      type: object
      properties:
        alias:
          type: string
          minLength: 1
          maxLength: 512
          description: User alias
        publicKey:
          type: string
          description: User public key
        password:
          type: string
          description: User password
    UserAuth:
      description: User auth data
      type: object
      properties:
        alias:
          type: string
          minLength: 1
          maxLength: 512
          description: User alias
        password:
          type: string
          description: User password
...